Add event listener and various functions
I add event listener for each buttons which call a function called
getTarget().

The getTarget() function determine the type of a button and pass it's
value to other functions which can store, display, delete, and clear
the value.

There are other functions to test different conditions and help the
above functions do their work correctly. these functions are
        - canConcatenate: test if a variable is null or not
        - noMoreDot: prevent adding dot if the value already has
        - concatenate: concatenate input to current(displayed) value
        - canEvaluate: test if operations need to be performed
        - StartEvaluate: start Mathematical operation
        - roundDecimals: round the result to 2 decimal place
        - resetValue: reset every value after operation has performed